+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.0.8] - 2023-05-22
+
+### Changed
+
+- Updated tasking to make sure specified parameter groups in the UI carry over
+- Updated tasking to list out unused parameters via the task's stdout/stderr modal
+
 ## [v1.0.7] - 2023-05-17
 
 ### Changed